+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+ - changed files signal to RwSignal and added extend option for `use_drop_zone`
+
 ## [0.13.13] - 2025-01-07
 
 ### New Functions 🚀